+++ runonce.sh
@@ -1,0 +1,3 @@
+cd rhtools-launcher && npm install
+cd electron/renderer && npm install
+